Moved Mobile detect to helpers, and removed address check
