+++ apps/site/assets/ts/helpers/mobileDetect.ts
