+++ apps/site/assets/ts/stop/components/StationInformation.tsx
@@ -10,8 +10,7 @@ const StationInformation = ({
   return (
     <div>
       <div>Station information PLACEHOLDER</div>
-      {/* TODO combine station name with municiplaity (station name, municiplaity) if no address */}
-      {stop.address && (
+      {
         <ExternalMapLink
           address={stop.address}
           municipality={stop.municipality}
@@ -19,7 +18,7 @@ const StationInformation = ({
           latitude={stop.latitude}
           longitude={stop.longitude}
         />
-      )}
+      }
       <div>Station Status Blocks PLACEHOLDER</div>
     </div>
   );
